wrap with on_load(:active_record) hook
